pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pr00t"
version = "0.1.0"
description = "Paras Robotics 00 Technology"
authors = [
    {name="Paras", email="parassakhare04@gmail.com"}
]
license = {text = "Apache 2.0"}
readme = "Readme.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.7.1",
    "torchvision",
    "torchaudio==2.7.1",
    "transformers",
    "diffusers==0.27.2",
    "accelerate",
    "opencv-python==4.9.0.80",
    "Pillow==10.3.0",
    "scikit-image==0.22.0",
    "numpy==1.26.4",
    "matplotlib==3.8.4",
    "tqdm==4.66.2",
    "scipy==1.13.0",
    "labelme==5.3.1",
    "imgaug==0.4.0",
    "jupyterlab==4.1.5",
    "ipykernel==6.29.4",
    "notebook==7.1.3",
    "easyocr==1.7.1",
    "networkx==3.2.1",
    "bitsandbytes==0.46.0"
]

[project.scripts]
pr00t = "pr00t.core:pr00t_main"

[tool.setuptools]
packages = ["pr00t"]